+++ src/onix/onix.js
@@ -694,6 +694,8 @@ onix = (function() {
 	/**
 	 * Framework info.
 	 *
+	 * version: 2.5.7
+	 * date: 2. 6. 2016
 	 * @member onix
 	 */
 	onix.info = function() {